Searchbot_TG
To use the bot, you'll need to create a Telegram bot using the BotFather and get its token. Then replace YOUR_TOKEN with your actual bot token in the code above.

Once the bot is running, users can send the command /price <TOKEN> to get the current price of a crypto token. For example, to get the price of Bitcoin, users can send the message /price bitcoin. The bot will then retrieve the name and current price of Bitcoin from the CoinGecko API and reply with a message like "Bitcoin: $59,123.45".